Adaption to python 3 and some internal updates on the whole algorithm...
