+++ multiply_atmospheric_corection/util/smoothn.py
@@ -174,23 +174,23 @@ def smoothn(y,nS0=10,axis=None,smoothOrder=2.0,sd=None,verbose=False,\
     mask = y.mask
     y = np.array(y)
     y[mask] = 0.
-    if W != None:
+    if np.any(W != None):
       W  = np.array(W)
       W[mask] = 0.
-    if sd != None:
+    if np.any(sd != None):
       W = np.array(1./sd**2)
       W[mask] = 0.
       sd = None
     y[mask] = np.nan
     
-  if sd != None:
+  if np.any(sd != None):
     sd_ = np.array(sd)
     mask = (sd > 0.)
     W = np.zeros_like(sd_)
     W[mask] = 1./sd_[mask]**2
     sd = None
 
-  if W != None:
+  if np.any(W != None):
     W = W/W.max()
 
   sizy = y.shape;
@@ -208,7 +208,7 @@ def smoothn(y,nS0=10,axis=None,smoothOrder=2.0,sd=None,verbose=False,\
   # Smoothness parameter and weights
   #if s != None:
   #  s = []
-  if W == None:
+  if np.all(W == None):
     W = ones(sizy);
 
   #if z0 == None:
@@ -256,11 +256,11 @@ def smoothn(y,nS0=10,axis=None,smoothOrder=2.0,sd=None,verbose=False,\
   Lambda = zeros(sizy);
   for i in axis:
     # create a 1 x d array (so e.g. [1,1] for a 2D case
-    siz0 = ones((1,y.ndim))[0];
+    siz0 = ones((1,y.ndim))[0].astype(int);
     siz0[i] = sizy[i];
     # cos(pi*(reshape(1:sizy(i),siz0)-1)/sizy(i)))
     # (arange(1,sizy[i]+1).reshape(siz0) - 1.)/sizy[i]
-    Lambda = Lambda + (cos(pi*(arange(1,sizy[i]+1) - 1.)/sizy[i]).reshape(siz0.astype(int)))
+    Lambda = Lambda + (cos(pi*(arange(1,sizy[i]+1) - 1.)/sizy[i]).reshape(siz0))
     #else:
     #  Lambda = Lambda + siz0
   Lambda = -2.*(len(axis)-Lambda);
@@ -331,7 +331,7 @@ def smoothn(y,nS0=10,axis=None,smoothOrder=2.0,sd=None,verbose=False,\
     #---
     while tol>TolZ and nit<MaxIter:
         if verbose:
-          print 'tol',tol,'nit',nit
+          print('tol',tol,'nit',nit)
         nit = nit+1;
         DCTy = dctND(Wtot*(y-z)+z,f=dct);
         if isauto and not remainder(log2(nit),1):
@@ -363,7 +363,7 @@ def smoothn(y,nS0=10,axis=None,smoothOrder=2.0,sd=None,verbose=False,\
               #print '==============='
             else:
               xpost = [s0]
-            xpost,f,d = lbfgsb.fmin_l_bfgs_b(gcv,xpost,fprime=None,factr=10.,\
+            xpost,f,d = lbfgsb.fmin_l_bfgs_b(gcv,xpost,fprime=None,factr=1e7,\
                approx_grad=True,bounds=[(log10(sMinBnd),log10(sMaxBnd))],\
                args=(Lambda,aow,DCTy,IsFinite,Wtot,y,nof,noe,smoothOrder))
         s = 10**xpost[0];
@@ -411,8 +411,8 @@ def smoothn(y,nS0=10,axis=None,smoothOrder=2.0,sd=None,verbose=False,\
   return z,s,exitflag,Wtot
 
 def warning(s1,s2):
-  print s1
-  print s2[0]
+  print(s1)
+  print(s2[0])
 
 ## GCV score
 #---
@@ -644,7 +644,7 @@ def sparseSVD(D):
   try:
     import sparsesvd
   except:
-    print 'bummer ... better get sparsesvd'
+    print('bummer ... better get sparsesvd')
     exit(0)
   Ds = scipy.sparse.csc_matrix(D)
   a = sparsesvd.sparsesvd(Ds,Ds.shape[0])
@@ -706,4 +706,5 @@ def sparseTest(n=1000):
   # Ut.T * Ut = I
   # ((Vt.T * (np.diag(np.array(eigenvalues).flatten())**2)) * Vt)
   # we see you get the same as m.T * m by squaring the eigenvalues
+
